Move store implementation to db package
