+++ backend/cmd/api/api.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"net/http"
 
+	"github.com/djhranicky/ConcertTracker-SE-Project/db"
 	"github.com/djhranicky/ConcertTracker-SE-Project/routes"
-	"github.com/djhranicky/ConcertTracker-SE-Project/service/user"
 	"github.com/gorilla/mux"
 	"gorm.io/gorm"
 )
@@ -26,7 +26,7 @@ func (s *APIServer) Run() error {
 	router := mux.NewRouter()
 	subrouter := router.PathPrefix("/api").Subrouter()
 
-	userStore := user.NewStore(s.db)
+	userStore := db.NewStore(s.db)
 	handler := routes.NewHandler(userStore)
 	handler.RegisterRoutes(subrouter)
