+++ backend/routes/routes_test.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/djhranicky/ConcertTracker-SE-Project/db"
 	"github.com/djhranicky/ConcertTracker-SE-Project/service/auth"
-	"github.com/djhranicky/ConcertTracker-SE-Project/service/user"
 	"github.com/djhranicky/ConcertTracker-SE-Project/types"
 	"github.com/djhranicky/ConcertTracker-SE-Project/utils"
 	"github.com/gorilla/mux"
@@ -339,7 +338,7 @@ func initTestDatabase(dbName string) *gorm.DB {
 
 func initTestHandler() (*Handler, *gorm.DB) {
 	database := initTestDatabase("test.db")
-	userStore := user.NewStore(database)
+	userStore := db.NewStore(database)
 	handler := NewHandler(userStore)
 
 	hashedPassword, err := auth.HashPassword("test")
